+++ walletkit/CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.20.0](https://github.com/worldcoin/walletkit/compare/v0.19.1...v0.20.0) - 2026-06-10
+
+### Other
+
+- introduce walletkit-db storage primitives crate ([#400](https://github.com/worldcoin/walletkit/pull/400))
+- V4.0 credential registration diagram ([#395](https://github.com/worldcoin/walletkit/pull/395))
+- make vault listener notification test deterministic ([#418](https://github.com/worldcoin/walletkit/pull/418))
+
 ## [0.19.1](https://github.com/worldcoin/walletkit/compare/v0.19.0...v0.19.1) - 2026-06-01
 
 ### Other